refactor: split scheduler modules
